+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * [UPGRADE] - upgraded to Angular 7
 * [NEW] - added toast for errors (error handler)
+* [NEW] - add http interceptor with error handling (offline error toast)
 
 ## v 4.1.1 (2019/02/26) (codename: maya-1)